Use sockets instead of Boost message queues for IPC
This is a lot more stable on the UI side.

+++ engine/src/modules/processing_graph/topology/anthem_graph_node.cpp
@@ -71,7 +71,7 @@ void AnthemGraphNode::initializePorts() {
   }
 }
 
-void AnthemGraphNode::setParameter(int id, float value) {
+void AnthemGraphNode::setParameter(uint64_t id, float value) {
   auto index = processor->config.getIndexOfParameter(id);
 
   if (!index.has_value()) return;
